Twilio Magic Values v0.2.0 TwilioMagicValues.BuyNumber

Wrapper for Twilio magic phone numbers used in tests buying phone numbers.

See: https://www.twilio.com/docs/iam/test-credentials#test-incoming-phone-numbers

Functions

available_area_code()

Returns the magic area code which has numbers available for purchase.

Expect no error.

empty_area_code()

Returns the magic area code which has NO numbers available for purchase.

Expect the 21452 API error from Twilio.

invalid()

Returns the magic phone number which is INVALID for purchase.

Expect the 21421 API error from Twilio.

unavailable()

Returns the magic phone number which is UNAVAILABLE for purchase.

Expect the 21421 API error from Twilio.

valid()

Returns the magic phone number which is VALID for purchase.

Expect no error.