Twilio Magic Values v0.3.0 TwilioMagicValues.Call

Wrapper for Twilio magic phone numbers used in tests making calls.

See: https://www.twilio.com/docs/iam/test-credentials#test-calls

Functions

blacklisted_recipient()

Returns the magic phone number which is BLACKLISTED to call to.

Expect the 21216 API error from Twilio.

Example

iex> map = make_call(from: TwilioMagicValues.Call.valid_sender(), to: TwilioMagicValues.Call.blacklisted_recipient())
iex> Map.get(map, :code)
21216

from_other(phone_number)

All other phone numbers, even ones owned by your production credentials, will return an error when used to call from.

(Other numbers are not verified on your test credentials.)

Expect the 21210 API error from Twilio.

Example

iex> map = make_call(from: TwilioMagicValues.Call.from_other("+17148675308"), to: TwilioMagicValues.Call.to_other("+17148675309"))
iex> Map.get(map, :code)
21210

geo_blocked_recipient()

Returns the magic phone number which has GEO PERMISSION ERRORS to call to.

Expect the 21215 API error from Twilio.

Example

iex> map = make_call(from: TwilioMagicValues.Call.valid_sender(), to: TwilioMagicValues.Call.geo_blocked_recipient())
iex> Map.get(map, :code)
21215

invalid_recipient()

Returns the magic phone number which is INVALID to call to.

Expect the 21217 API error from Twilio.

Example

iex> map = make_call(from: TwilioMagicValues.Call.valid_sender(), to: TwilioMagicValues.Call.invalid_recipient())
iex> Map.get(map, :code)
21217

invalid_sender()

Returns the magic phone number which is INVALID to call from.

Expect the 21212 API error from Twilio.

Example

iex> map = make_call(from: TwilioMagicValues.Call.invalid_sender(), to: TwilioMagicValues.Call.to_other("+17148675309"))
iex> Map.get(map, :code)
21212

to_other(phone_number)

All other recipient phone numbers are potentially valid and are verified normally.

Errors are dependent on input.

unroutable_recipient()

Returns the magic phone number which is UNROUTABLE to call to.

Expect the 21214 API error from Twilio.

Example

iex> map = make_call(from: TwilioMagicValues.Call.valid_sender(), to: TwilioMagicValues.Call.unroutable_recipient())
iex> Map.get(map, :code)
21214

valid_sender()

Returns the magic phone number which is VALID to call from.

Expect no error.

Example

iex> map = make_call(from: TwilioMagicValues.Call.valid_sender(), to: TwilioMagicValues.Call.to_other("+17148675309"))
iex> Map.get(map, :code, "There is no error code")
"There is no error code"
iex> Map.get(map, :from_formatted)
"(500) 555-0006"