Twilio Magic Values v0.3.0 TwilioMagicValues.Message

Wrapper for Twilio magic phone numbers used in tests sending SMS messages.

See: https://www.twilio.com/docs/iam/test-credentials#test-sms-messages

Functions

blacklisted_recipient()

Returns the magic phone number which is BLACKLISTED to send a message to.

Expect the 21610 API error from Twilio.

cant_receive_sms()

Returns the magic phone number which is DOESNT SUPPORT SMS to send a message to.

Expect the 21614 API error from Twilio.

geo_blocked_recipient()

Returns the magic phone number which has a GEO PERMISSION ERROR to send a message to.

Expect the 21408 API error from Twilio.

invalid_recipient()

Returns the magic phone number which is INVALID to send a message to.

Expect the 21211 API error from Twilio.

invalid_sender()

Returns the magic phone number which is INVALID to send SMS from.

Expect the 21212 API error from Twilio.

sender_cant_send_sms()

Returns the magic phone number which is CANT SEND SMS to send SMS from.

Expect the 21606 API error from Twilio.

sender_queue_full()

Returns the magic phone number which has A FULL QUEUE to send SMS from.

Expect the 21611 API error from Twilio.

to_other(phone_number)

All other phone numbers are potentially valid and are verified normally.

Errors are dependent on input.

unowned_number()

Returns the magic phone number which is NOT VERIFIED BY YOUR ACCOUNT to send SMS from.

Expect the 21606 API error from Twilio.

unroutable_recipient()

Returns the magic phone number which is UNROUTABLE to send a message to.

Expect the 21612 API error from Twilio.

valid_sender()

Returns the magic phone number which is VALID to send SMS from.

Expect no error.